Move artifact contracts out of core index
