+++ packages/runtime-core/src/artifact-test-results.ts
@@ -1,0 +1,29 @@
+export interface ArtifactTestResultsRawLogReference {
+  path: string
+  kind: string
+}
+
+export interface ArtifactTestResultsSuite {
+  name: string
+  status: "passed" | "failed" | "skipped" | "unknown"
+  tests: number
+  passed: number
+  failed: number
+  skipped: number
+  unknown: number
+  rawLogReferences?: ArtifactTestResultsRawLogReference[]
+}
+
+export interface ArtifactTestResults {
+  schema: "wp-codebox/test-results/v1"
+  status: "passed" | "failed" | "skipped" | "unknown"
+  summary: {
+    total: number
+    passed: number
+    failed: number
+    skipped: number
+    unknown: number
+  }
+  suites: ArtifactTestResultsSuite[]
+  rawLogReferences: ArtifactTestResultsRawLogReference[]
+}
