+++ tests/fixtures/artifact-package-provenance-shape.json
@@ -2,15 +2,15 @@
   "schema": "wp-codebox/package-provenance/v1",
   "wpCodebox": {
     "name": "wp-codebox-workspace",
-    "version": "0.7.2"
+    "version": "0.7.3"
   },
   "runtimeCore": {
     "name": "@automattic/wp-codebox-core",
-    "version": "0.7.2"
+    "version": "0.7.3"
   },
   "runtimePlayground": {
     "name": "@automattic/wp-codebox-playground",
-    "version": "0.7.2"
+    "version": "0.7.3"
   },
   "playground": {
     "cli": {
